회원가입 기능, 맞는 유저 글 수정 삭제, 관리자 글 수정 삭제

+++ pages/api/delete/deleteItem.ts
@@ -1,13 +1,25 @@
 import { ServerPropsType } from "@/src/type/types";
 import { connectDB } from "@/src/util/database";
 import { ObjectId } from "mongodb";
+import { getServerSession } from "next-auth";
+import { authOptions } from "../auth/[...nextauth]";
 
 const handler = async (...[req, res]: ServerPropsType) => {
   if (req.method === "DELETE") {
     try {
+      const session = await getServerSession(req, res, authOptions);
+
       const db = (await connectDB).db("hellostory");
-      const result = db.collection("post").deleteOne({ _id: new ObjectId(req.body) });
-      return res.status(200).json("success");
+      const findItme = await db.collection("post").findOne({ _id: new ObjectId(req.body) });
+      // if (findItme?.author !== session.user.name) {
+      //   return res.status(403).json("접근 권한 없음");
+      // }
+      if (findItme?.author === session.user.name || session.user.role === "admin") {
+        const result = db.collection("post").deleteOne({ _id: new ObjectId(req.body) });
+        return res.status(200).json("success");
+      } else {
+        return res.status(403).json("접근 권한 없음");
+      }
     } catch (error) {
       return res.status(500).json("DB에러");
     }
